requirements.txt
azure-functions
azure-identity
azure-keyvault-secrets
azure-keyvault-keys
azure-keyvault-certificates
